# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/project/{projectId}/versions/{versionId}/feedbacks": {
      "post": {
        "operationId": "FeedbackController_createFeedback",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "versionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFeedbackDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateFeedbackResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Feedback"
        ]
      },
      "get": {
        "operationId": "FeedbackController_findFeedbacksForVersion",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "versionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Feedback"
        ]
      }
    },
    "/project/{projectId}/versions/{versionId}/feedbackQuestions": {
      "get": {
        "operationId": "FeedbackController_findAllQuestions",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "versionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackQuestionsResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Feedback"
        ]
      }
    },
    "/project/{projectId}/versions/{versionId}/feedbacks/{submissionId}/unlock": {
      "post": {
        "operationId": "FeedbackController_unlockFeedback",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "versionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "submissionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnlockFeedbackDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnlockFeedbackResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Feedback"
        ]
      }
    },
    "/feedbacks/recent": {
      "get": {
        "operationId": "MyFeedbackController_getRecentFeedbacks",
        "parameters": [
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 30,
              "default": 6,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecentFeedbacksResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Feedback"
        ]
      }
    },
    "/feedbacks/my/projects": {
      "get": {
        "operationId": "MyFeedbackController_findMyFeedbackProjects",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyFeedbackProjectsResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Feedback"
        ]
      }
    },
    "/feedbacks/{submissionId}": {
      "get": {
        "operationId": "MyFeedbackController_findFeedbackSubmissionDetail",
        "parameters": [
          {
            "name": "submissionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackSubmissionDetailResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Feedback"
        ]
      }
    },
    "/project/{projectId}/feedbacks": {
      "post": {
        "operationId": "ProjectFeedbackController_createFreeformFeedback",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFreeformFeedbackDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateFeedbackResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Feedback"
        ]
      },
      "get": {
        "operationId": "ProjectFeedbackController_findFreeformFeedbacks",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Feedback"
        ]
      }
    },
    "/project/{projectId}/feedbacks/{submissionId}/unlock": {
      "post": {
        "operationId": "ProjectFeedbackController_unlockFreeformFeedback",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "submissionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnlockFeedbackDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnlockFeedbackResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Feedback"
        ]
      }
    },
...
    "/project/{id}": {
      "get": {
        "operationId": "ProjectController_getProjectById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetailResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Project"
        ]
      },
      "patch": {
        "operationId": "ProjectController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Project"
        ]
      }
    },
...
    "/project/{id}/versions/{versionId}": {
      "get": {
        "operationId": "GrowthRecordController_getVersionDetail",
        "parameters": [
          {
            "name": "versionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionDetailResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "GrowthRecord"
        ]
      }
    },
...
  },
  "info": {
    "title": "Proseed API",
    "description": "사이드 프로젝트 성장기록·피드백 플랫폼 API. 인증은 better-auth 세션 쿠키 기반이며, 로그인 라우트(/api/auth/*)는 better-auth가 별도로 처리하므로 이 문서에 포함되지 않는다. 로컬에서는 ENABLE_DEV_LOGIN=true 후 POST /api/auth/sign-in/email (seed 계정)으로 세션을 얻을 수 있다.",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token"
      }
    },
    "schemas": {
...
      "MyFeedbackProjectsResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MyFeedbackProjectItemDto"
            }
          }
        },
        "required": [
          "success",
          "data"
        ]
      },
      "FeedbackItemDto": {
        "type": "object",
        "properties": {
          "questionId": {
            "type": "number"
          },
          "content": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "imageUrls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "questionId",
          "content"
        ]
      },
      "CreateFeedbackDto": {
        "type": "object",
        "properties": {
          "oneLineReview": {
            "type": "string"
          },
          "feedbacks": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FeedbackItemDto"
            }
          }
        },
        "required": [
          "oneLineReview",
          "feedbacks"
        ]
      },
...
      "CreateFeedbackDataDto": {
        "type": "object",
        "properties": {
          "submittedCount": {
            "type": "number"
          },
          "feedbacks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreatedFeedbackDto"
            }
          }
        },
        "required": [
          "submittedCount",
          "feedbacks"
        ]
      },
      "CreateFeedbackResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/CreateFeedbackDataDto"
          }
        },
        "required": [
          "success",
          "data"
        ]
      },
      "RecordCategory": {
        "type": "string",
        "enum": [
          "PLAN",
          "DESIGN",
          "DEVELOPMENT",
          "GENERAL"
        ]
      },
      "FeedbackQuestionItemDto": {
        "type": "object",
        "properties": {
          "category": {
            "allOf": [
              {
                "$ref": "#/components/schemas/RecordCategory"
              }
            ]
          },
          "id": {
            "type": "number"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "order": {
            "type": "number"
          },
          "required": {
            "type": "boolean"
          }
        },
        "required": [
          "category",
          "id",
          "title",
          "description",
          "order",
          "required"
        ]
      },
      "FeedbackQuestionsResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FeedbackQuestionItemDto"
            }
          }
        },
        "required": [
          "success",
          "data"
        ]
      },
...
      "FeedbackListItemDto": {
        "type": "object",
        "properties": {
          "category": {
            "allOf": [
              {
                "$ref": "#/components/schemas/RecordCategory"
              }
            ]
          },
          "imageUrls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "submissionId": {
            "type": "number"
          },
          "userId": {
            "type": "number"
          },
          "author": {
            "$ref": "#/components/schemas/FeedbackSubmissionAuthorDto"
          },
          "oneLineReview": {
            "type": "string"
          },
          "isAdopted": {
            "type": "boolean",
            "description": "이미 다른 성장기록에 채택(태그)된 제출인지 — 성장기록 발행의 피드백 태그하기에서 재태그 방지용"
          },
          "isUnlocked": {
            "type": "boolean",
            "description": "이 제출이 열람(unlock)됐는지. false면 content/imageUrls는 서버에서 비워서 내려감"
          },
          "id": {
            "type": "number"
          },
          "questionId": {
            "type": "number",
            "description": "성장기록 없이 남긴 자유 피드백은 실제 질문이 없으므로 0"
          },
          "questionTitle": {
            "type": "string"
          },
          "questionContent": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "category",
          "imageUrls",
          "submissionId",
          "userId",
          "author",
          "oneLineReview",
          "isAdopted",
          "isUnlocked",
          "id",
          "questionId",
          "questionTitle",
          "questionContent",
          "content",
          "createdAt",
          "updatedAt"
        ]
      },
      "FeedbackListResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FeedbackListItemDto"
            }
          }
        },
        "required": [
          "success",
          "data"
        ]
      },
      "UnlockFeedbackDto": {
        "type": "object",
        "properties": {
          "category": {
            "allOf": [
              {
                "$ref": "#/components/schemas/RecordCategory"
              }
            ]
          }
        },
        "required": [
          "category"
        ]
      },
      "UnlockFeedbackDataDto": {
        "type": "object",
        "properties": {
          "category": {
            "allOf": [
              {
                "$ref": "#/components/schemas/RecordCategory"
              }
            ]
          },
          "submissionId": {
            "type": "number"
          },
          "isUnlocked": {
            "type": "boolean"
          },
          "charged": {
            "type": "boolean",
            "description": "이번 요청에서 실제로 티켓이 차감됐는지. 이미 열려 있던 제출이면 false(무과금)"
          },
          "remainingTickets": {
            "type": "number",
            "description": "차감 후 남은 티켓 수"
          }
        },
        "required": [
          "category",
          "submissionId",
          "isUnlocked",
          "charged",
          "remainingTickets"
        ]
      },
      "UnlockFeedbackResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/UnlockFeedbackDataDto"
          }
        },
        "required": [
          "success",
          "data"
        ]
      },
      "RecentFeedbackItemDto": {
        "type": "object",
        "properties": {
          "category": {
            "allOf": [
              {
                "$ref": "#/components/schemas/RecordCategory"
              }
            ]
          },
          "submissionId": {
            "type": "number"
          },
          "versionId": {
            "type": "number"
          },
          "nickname": {
            "type": "string"
          },
          "profileImageUrl": {
            "type": "string"
          },
          "oneLineReview": {
            "type": "string"
          },
          "content": {
            "type": "string",
            "description": "채택된 직군 질문에 대한 첫 번째 답변 본문"
          },
          "projectId": {
            "type": "number"
          },
          "projectName": {
            "type": "string"
          },
          "projectIconUrl": {
            "type": "string",
            "description": "presigned download URL"
          }
        },
        "required": [
          "category",
          "submissionId",
          "versionId",
          "nickname",
          "profileImageUrl",
          "oneLineReview",
          "content",
          "projectId",
          "projectName",
          "projectIconUrl"
        ]
      },
      "RecentFeedbacksResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecentFeedbackItemDto"
            }
          }
        },
        "required": [
          "success",
          "data"
        ]
      },
...
      "FeedbackSubmissionDetailResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/FeedbackSubmissionDetailDto"
          }
        },
        "required": [
          "success",
          "data"
        ]
      },
...
      "CreateFreeformFeedbackDto": {
        "type": "object",
        "properties": {
          "oneLineReview": {
            "type": "string"
          },
          "feedbacks": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FreeformFeedbackItemDto"
            }
          }
        },
        "required": [
          "oneLineReview",
          "feedbacks"
        ]
      },
...
      "ProjectType": {
        "type": "string",
        "enum": [
          "APP",
          "WEB"
        ]
      },
...
      "ProjectResponseDto": {
        "type": "object",
        "properties": {
          "type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ProjectType"
              }
            ]
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ProjectStatus"
              }
            ]
          },
          "category": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectCategory"
            }
          },
          "id": {
            "type": "number"
          },
          "title": {
            "type": "string"
          },
          "createdById": {
            "type": "number"
          },
          "oneLineDescription": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "contactPath": {
            "type": "string"
          },
          "projectLink": {
            "type": "string"
          },
          "iconUrl": {
            "type": "string"
          },
          "thumbnailUrl": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "type",
          "status",
          "category",
          "id",
          "title",
          "createdById",
          "oneLineDescription",
          "description",
          "contactPath",
          "projectLink",
          "iconUrl",
          "thumbnailUrl",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "ProjectDetailResponseDto": {
        "type": "object",
        "properties": {
          "type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ProjectType"
              }
            ]
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ProjectStatus"
              }
            ]
          },
          "category": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectCategory"
            }
          },
          "myJobType": {
            "nullable": true,
            "description": "요청자가 참여한 직군 (비로그인·비참여 시 null). 성장기록은 이 직군만 작성 가능",
            "allOf": [
              {
                "$ref": "#/components/schemas/JobType"
              }
            ]
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectImageDto"
            }
          },
          "projectRoles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectMemberDto"
            }
          },
          "isMyProject": {
            "type": "boolean",
            "description": "요청자가 이 프로젝트의 멤버인지 (비로그인 시 false)"
          },
          "id": {
            "type": "number"
          },
          "title": {
            "type": "string"
          },
          "createdById": {
            "type": "number"
          },
          "oneLineDescription": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "contactPath": {
            "type": "string"
          },
          "projectLink": {
            "type": "string"
          },
          "iconUrl": {
            "type": "string"
          },
          "thumbnailUrl": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "type",
          "status",
          "category",
          "myJobType",
          "images",
          "projectRoles",
          "isMyProject",
          "id",
          "title",
          "createdById",
          "oneLineDescription",
          "description",
          "contactPath",
          "projectLink",
          "iconUrl",
          "thumbnailUrl",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateProjectDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "type": {
            "enum": [
              "APP",
              "WEB"
            ],
            "type": "string"
          },
          "status": {
            "enum": [
              "Available",
              "MVP",
              "Ongoing",
              "Hiring"
            ],
            "type": "string"
          },
          "oneLineDescription": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "category": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "HEALTHCARE",
                "FINANCE",
                "PUBLIC",
                "COMMERCE",
                "EDUCATION",
                "ENTERTAINMENT",
                "MOBILITY",
                "ENERGY",
                "REALESTATE",
                "LIFESTYLE",
                "PRODUCTIVITY",
                "COMMUNITY",
                "AI"
              ]
            }
          },
          "contactPath": {
            "type": "string"
          },
          "projectLink": {
            "type": "string"
          },
          "iconKey": {
            "type": "string"
          },
          "thumbnailKey": {
            "type": "string"
          },
          "imageKeys": {
            "maxItems": 8,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "VersionDetailResponseDto": {
        "type": "object",
        "properties": {
          "releasedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "number"
          },
          "projectId": {
            "type": "number"
          },
          "version": {
            "type": "string"
          },
          "updateGoal": {
            "type": "string"
          },
          "updateResults": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "growthRecords": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VersionDetailGrowthRecordDto"
            }
          },
          "feedbackQuestions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VersionFeedbackQuestionDto"
            }
          }
        },
        "required": [
          "releasedAt",
          "id",
          "projectId",
          "version",
          "updateGoal",
          "updateResults",
          "createdAt",
          "updatedAt",
          "growthRecords",
          "feedbackQuestions"
        ]
      },
...
    }
  }
}